+++ build_runner/test/build_script_generate/build_script_generate_test.dart
@@ -4,8 +4,6 @@
 @Timeout.factor(4)
 import 'package:_test_common/descriptors.dart';
 import 'package:_test_common/sdk.dart';
-import 'package:build_runner/src/build_script_generate/build_script_generate.dart';
-import 'package:build_runner_core/build_runner_core.dart';
 import 'package:test/test.dart';
 import 'package:test_descriptor/test_descriptor.dart' as d;
 
@@ -39,7 +37,8 @@ builders:
         ]).create();
 
         var result = await runPub('a', 'run', args: ['build_runner', 'build']);
-        expect(result.stderr, isEmpty);
+        expect(result.stderr, isEmpty,
+            skip: 'https://github.com/dart-lang/sdk/issues/50592');
         expect(
             result.stdout,
             contains(
@@ -58,7 +57,8 @@ builders:
         ]).create();
 
         var result = await runPub('a', 'run', args: ['build_runner', 'build']);
-        expect(result.stderr, isEmpty);
+        expect(result.stderr, isEmpty,
+            skip: 'https://github.com/dart-lang/sdk/issues/50592');
         expect(
             result.stdout,
             isNot(contains(
@@ -87,8 +87,28 @@ builders:
 ''')
         ]).create();
         var result = await runPub('a', 'run', args: ['build_runner', 'build']);
-        expect(result.stderr, isEmpty);
+        expect(result.stderr, isEmpty,
+            skip: 'https://github.com/dart-lang/sdk/issues/50592');
         expect(result.stdout, contains('could not be parsed'));
+      });
+
+      test('warn when import not present in packageGraph', () async {
+        await d.dir('a', [
+          d.file('build.yaml', '''
+builders:
+  fake:
+    import: "package:unknown_package/import.dart"
+    builder_factories: ["myFactory"]
+    build_extensions: {"foo": ["bar"]}
+''')
+        ]).create();
+        var result = await runPub('a', 'run', args: ['build_runner', 'build']);
+        expect(result.stderr, isEmpty,
+            skip: 'https://github.com/dart-lang/sdk/issues/50592');
+        expect(
+            result.stdout,
+            contains('Could not load imported package "unknown_package" '
+                'for definition "a:fake".'));
       });
     });
 
@@ -103,7 +123,8 @@ global_options:
       ]).create();
 
       var result = await runPub('a', 'run', args: ['build_runner', 'build']);
-      expect(result.stderr, isEmpty);
+      expect(result.stderr, isEmpty,
+          skip: 'https://github.com/dart-lang/sdk/issues/50592');
       expect(
           result.stdout,
           allOf(
@@ -114,143 +135,6 @@ global_options:
                 'Invalid builder key `b:b` found in global_options config of '
                 'build.yaml. This configuration will have no effect.'),
           ));
-    });
-  });
-
-  group('determines whether a build script can be null safe', () {
-    final buildYaml = d.file('build.yaml', '''
-builders:
-  a:
-    import: 'package:a/builder.dart'
-    build_extensions: {'.foo': ['.bar']}
-    builder_factories: ['builder']
-    ''');
-
-    test('when all builders are opted-in', () async {
-      await d.dir('a', [
-        d.file('pubspec.yaml', '''
-name: a
-environment:
-  sdk: '>=2.12.0 <3.0.0'
-      '''),
-        buildYaml,
-        d.dir('lib', [d.file('builder.dart', '')]),
-      ]).create();
-      await runPub('a', 'get');
-
-      final options = await findBuildScriptOptions(
-          packageGraph: await PackageGraph.forPath('${d.sandbox}/a'));
-      expect(options.canRunWithSoundNullSafety, isTrue);
-    });
-
-    test('when the root package opts out', () async {
-      await d.dir('a', [
-        d.file('pubspec.yaml', '''
-name: a
-environment:
-  sdk: '>=2.9.0 <3.0.0'
-      '''),
-        d.dir('lib', [d.file('builder.dart', '')]),
-      ]).create();
-      await runPub('a', 'get');
-
-      final options = await findBuildScriptOptions(
-          packageGraph: await PackageGraph.forPath('${d.sandbox}/a'));
-      expect(options.canRunWithSoundNullSafety, isFalse);
-    });
-
-    test('when a builder-defining package opts out', () async {
-      await d.dir('a', [
-        d.file('pubspec.yaml', '''
-name: a
-environment:
-  sdk: '>=2.12.0 <3.0.0'
-dependencies:
-  b:
-    path: ../b/
-      '''),
-      ]).create();
-      await d.dir('b', [
-        d.file('pubspec.yaml', '''
-name: b
-environment:
-  sdk: '>=2.9.0 <3.0.0'
-      '''),
-        buildYaml,
-        d.dir('lib', [
-          d.file('builder.dart', ''),
-        ]),
-      ]).create();
-      await runPub('a', 'get');
-
-      final options = await findBuildScriptOptions(
-          packageGraph: await PackageGraph.forPath('${d.sandbox}/a'));
-      expect(options.canRunWithSoundNullSafety, isFalse);
-    });
-
-    test('when a builder-defining library ops out', () async {
-      await d.dir('a', [
-        d.file('pubspec.yaml', '''
-name: a
-environment:
-  sdk: '>=2.12.0 <3.0.0'
-      '''),
-        buildYaml,
-        d.dir('lib', [d.file('builder.dart', '// @dart=2.9')]),
-      ]).create();
-      await runPub('a', 'get');
-
-      final options = await findBuildScriptOptions(
-          packageGraph: await PackageGraph.forPath('${d.sandbox}/a'));
-      expect(options.canRunWithSoundNullSafety, isFalse);
-    });
-
-    test('when a builder-defining library ops out through a relative path',
-        () async {
-      await d.dir('a', [
-        d.file('pubspec.yaml', '''
-name: a
-environment:
-  sdk: '>=2.12.0 <3.0.0'
-      '''),
-        d.file('build.yaml', '''
-builders:
-  a:
-    import: '../../../tool/builder.dart'
-    build_extensions: {'.foo': ['.bar']}
-    builder_factories: ['builder']
-    '''),
-        d.dir('tool', [d.file('builder.dart', '//@dart=2.9')]),
-      ]).create();
-      await runPub('a', 'get');
-
-      final options = await findBuildScriptOptions(
-          packageGraph: await PackageGraph.forPath('${d.sandbox}/a'));
-      expect(options.canRunWithSoundNullSafety, isFalse);
-    });
-
-    test('when a builder-defining library uses a top-level relative path',
-        () async {
-      await d.dir('a', [
-        d.file('pubspec.yaml', '''
-name: a
-environment:
-  sdk: '>=2.12.0 <3.0.0'
-      '''),
-        d.file('build.yaml', '''
-builders:
-  a:
-    import: 'tool/builder.dart'
-    build_extensions: {'.foo': ['.bar']}
-    builder_factories: ['builder']
-    '''),
-        d.dir('tool', [d.file('builder.dart', '')]),
-      ]).create();
-      await runPub('a', 'get');
-
-      final options = await findBuildScriptOptions(
-          packageGraph: await PackageGraph.forPath('${d.sandbox}/a'));
-      expect(options.canRunWithSoundNullSafety, isTrue);
     });
   });
 }