+++ build_web_compilers/test/util.dart
@@ -6,12 +6,15 @@ import 'dart:async';
 
 import 'package:build/build.dart';
 import 'package:build_test/build_test.dart';
+import 'package:test/test.dart';
 
 /// Forwards to [testBuilder], and adds all output assets to [assets].
 Future<void> testBuilderAndCollectAssets(
     Builder builder, Map<String, Object> assets) async {
   var writer = InMemoryAssetWriter();
-  await testBuilder(builder, assets, writer: writer);
+  await testBuilder(builder, assets,
+      writer: writer,
+      onLog: (log) => printOnFailure('${log.level}: ${log.message}'));
   writer.assets.forEach((id, value) {
     assets['${id.package}|${id.path}'] = value;
   });